Correct softwareReset startup delay
TN017 has its own separate start-up delay of 2ms for software reset procedures that acts independent of the KXTJ3-1057 start-up delay.

The start-up delay variation in the KXTJ3-1057 datasheet is specific to setting the PC1 bit of CTRL_REG1 to 1. It is unrelated to setting the SRST bit of CTRL_REG2 to 1, as this is covered in TN017 instead.