Add software reset functionality (#7)
* Add softwareReset, fix formatting

Adds softwareReset function to reset the IMU according to Kionix TN017, performs softwareReset on begin() per TN017, makes start-up delay explicitly performed following exit of stand-by mode, and fixes some formatting and whitespace issues.

* Update spacing

Updates spacing in cpp

* More spacing updates

Fixed further spacing problems

* Whitespace fix

* Header whitespace changes

* Correct softwareReset startup delay

TN017 has its own separate start-up delay of 2ms for software reset procedures that acts independent of the KXTJ3-1057 start-up delay.

The start-up delay variation in the KXTJ3-1057 datasheet is specific to setting the PC1 bit of CTRL_REG1 to 1. It is unrelated to setting the SRST bit of CTRL_REG2 to 1, as this is covered in TN017 instead.

* Fix errors

* More fixes

* clang format

* run runner on pull request

* startupDelay do not need to be public

---------

Co-authored-by: ldab