Remove serial parameter from begin()
Due to the Adafruit Trinket M0 using an emulated serial port over USB defined as Serial_ instead of HardwareSerial like on most Arduinos, we cannot pass references or pointers to HardwareSerial while still successfully compiling on the M0. As such, the ability to define a serial port in the sketch has been removed and replaced with a define in the CPP file that can be changed by the user.