Interrupt, Debug, Example updates (#11)
* Interrupt routine fixes

Corrects INT_CTRL_REG1 routine to no longer set STPOL as that bit is only for enabling self-test, corrects bitshift for INT pin polarity, and adds routine to match the Wake-Up Data Rate as closely as possible to the user-set Output Data Rate.

* Fix debug, update example, add comments

Added comments to WUODR to make it obvious which data rate was being set in the code, fixed KTXJ3_DEBUG definition so that it will actually use the serial port passed to that definition by the user, updated basic example sketch to remove interrupt as it is not basic to use and did not work with the sketch as-written, and also corrected if statements and comments

* Correct clang errors

c l a n g c l a n g c l a n g

* Clang why

Why is clang like this it makes no sense

* Update README.md

Updated README to clarify that the maximum ODR for Interrupts is limited to 100 Hz regardless of the IMU's maximum ODR.

* Fix example bug

Example sketch was checking to see if IMU_SUCCESS was false rather than true in loop().

* Fix high resolution & debug modes

Previous implementation using preprocessor definitions was failing due to incorrect context switching between the CPP file and the sketch. These definitions have now been replaced with parameters passed to begin() in order to enable High Resolution mode or Debug mode, as well as set which Serial port to use for debugging in case this is being used on chipsets supporting more than one Serial port.

* Correct compiler errors

Previous iteration worked on certain cores but not others. As such, have changed the Serial definition to be a pointer, which should be supported on all cores.

* Remove serial parameter from begin()

Due to the Adafruit Trinket M0 using an emulated serial port over USB defined as Serial_ instead of HardwareSerial like on most Arduinos, we cannot pass references or pointers to HardwareSerial while still successfully compiling on the M0. As such, the ability to define a serial port in the sketch has been removed and replaced with a define in the CPP file that can be changed by the user.